backend/requirements.txt
# Core dependencies (Phase 0-1)
setuptools>=65.5.0
wheel
Flask==3.0.0
flask-cors==4.0.0
gunicorn==21.2.0

# Database (Phase 1)
psycopg2-binary==2.9.9
SQLAlchemy==2.0.23
python-dotenv==1.0.0
pgvector==0.2.4

# Computer Vision - Face Recognition (Phase 1)
opencv-python==4.8.1.78
numpy==1.24.3
face_recognition==1.3.0
scikit-learn==1.3.2
Pillow==10.1.0

# Phase 2 - Vision Intelligence Dependencies

# Emotion Detection
deepface==0.0.79
tf-keras==2.16.0  # Required by DeepFace

# Object Detection
ultralytics==8.0.220  # YOLOv8
torch==2.1.1  # Required by YOLO and CLIP
torchvision==0.16.1

# CLIP Embeddings
transformers==4.35.2
sentencepiece==0.1.99  # Required by some transformers models

# Additional utilities
tqdm==4.66.1  # Progress bars
python-dateutil==2.8.2  # Date parsing

# Optional: GPU support (uncomment if you have CUDA)
# torch==2.1.1+cu118
# torchvision==0.16.1+cu118